# Backend/api/laravel-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proyecto Final Grado API",
        "description": "Documentacion de la API para el Proyecto Final de Grado",
        "contact": {
            "name": "Alejandro Solar, David Kovacs",
            "email": "alejandrosolarruiz@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/generos": {
            "get": {
                "tags": [
                    "Genero"
                ],
                "summary": "Listar todos los géneros paginados",
                "operationId": "3f481aeb2388ba232907c09ac43e7f7d",
                "parameters": [
                    {
                        "name": "pagina",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "registrosPorPagina",
                        "in": "query",
                        "description": "Cantidad de registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de géneros paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Genero"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Genero"
                ],
                "summary": "Crear un nuevo género",
                "operationId": "e79f2f2d12fe7a22ed1d0754903b3c9d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "imagen": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "imagen": {
                                        "description": "Imagen en base64 (opcional)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Género creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genero"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/juegos": {
            "get": {
                "tags": [
                    "Juego"
                ],
                "summary": "Listar todos los juegos con sus imágenes y categorías (paginado, solo listado)",
                "operationId": "e6414dd14d51162ed1d94054ea30fe65",
                "parameters": [
                    {
                        "name": "pagina",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "registrosPorPagina",
                        "in": "query",
                        "description": "Cantidad de registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de juegos con imágenes (solo listado, sin metadatos de paginación)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "allOf": [
                                            {
                                                "$ref": "#/components/schemas/Juego"
                                            },
                                            {
                                                "properties": {
                                                    "juego_imagens": {
                                                        "type": "array",
                                                        "items": {
                                                            "$ref": "#/components/schemas/JuegoImagen"
                                                        }
                                                    },
                                                    "generos": {
                                                        "type": "array",
                                                        "items": {
                                                            "$ref": "#/components/schemas/Genero"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Juego"
                ],
                "summary": "Crear un nuevo juego",
                "operationId": "04fab3ea12216b2984f5f7fce0e35d58",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JuegoCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Juego creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Juego"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/juegos/{id}": {
            "get": {
                "tags": [
                    "Juego"
                ],
                "summary": "Obtener un juego por ID con sus imágenes y categorías",
                "operationId": "7330d442c22044dce5c0cf1a2c5a6dde",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del juego",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Juego encontrado con imágenes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Juego"
                                        },
                                        {
                                            "properties": {
                                                "juego_imagens": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/JuegoImagen"
                                                    }
                                                },
                                                "generos": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Genero"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Juego no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Juego"
                ],
                "summary": "Actualizar un juego existente",
                "operationId": "844e300d0024b08c77ebe846acd00ee1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del juego",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JuegoUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Juego actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Juego"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Juego no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Juego"
                ],
                "summary": "Eliminar un juego",
                "operationId": "0fc56b3e97500af5eacc5db163658a3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del juego",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Juego eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Juego eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Juego no encontrado"
                    }
                }
            }
        },
...
        "/api/juegos/buscarPorParametro": {
            "get": {
                "tags": [
                    "Juego"
                ],
                "summary": "Buscar juegos por nombre o género",
                "operationId": "36796057b56140494db41911822f5962",
                "parameters": [
                    {
                        "name": "parametro",
                        "in": "query",
                        "description": "Texto a buscar en el nombre del juego o en el nombre del género",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de juegos encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Juego"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/juegos/filtrar": {
            "get": {
                "tags": [
                    "Juego"
                ],
                "summary": "Filtrar juegos por nombre, género y/o plataforma",
                "operationId": "f8e4a22ececeadb2c61bb47ad4447f8d",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Filtrar por nombre del juego",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "genero",
                        "in": "query",
                        "description": "Filtrar por género",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "plataforma",
                        "in": "query",
                        "description": "Filtrar por plataforma",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de juegos filtrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Juego"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Genero": {
                "properties": {
                    "id_genero": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Acción"
                    }
                },
                "type": "object"
            },
...
            "Juego": {
                "properties": {
                    "id_juego": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "CyberQuest"
                    },
                    "desarrollador": {
                        "type": "string",
                        "example": "TechSoft"
                    },
                    "editor": {
                        "type": "string",
                        "example": "GameHouse"
                    },
                    "fecha_lanzamiento": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-06-15"
                    },
                    "precio": {
                        "type": "number",
                        "format": "float",
                        "example": 49.99
                    },
                    "descuento": {
                        "type": "number",
                        "format": "float",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "JuegoCreate": {
                "required": [
                    "nombre",
                    "precio"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "CyberQuest"
                    },
                    "desarrollador": {
                        "type": "string",
                        "example": "TechSoft"
                    },
                    "editor": {
                        "type": "string",
                        "example": "GameHouse"
                    },
                    "fecha_lanzamiento": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-06-15"
                    },
                    "precio": {
                        "type": "number",
                        "format": "float",
                        "example": 49.99
                    },
                    "descuento": {
                        "type": "number",
                        "format": "float",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "JuegoUpdate": {
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "CyberQuest"
                    },
                    "desarrollador": {
                        "type": "string",
                        "example": "TechSoft"
                    },
                    "editor": {
                        "type": "string",
                        "example": "GameHouse"
                    },
                    "fecha_lanzamiento": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-06-15"
                    },
                    "precio": {
                        "type": "number",
                        "format": "float",
                        "example": 49.99
                    },
                    "descuento": {
                        "type": "number",
                        "format": "float",
                        "example": 10
                    }
                },
                "type": "object"
            },
...
            "JuegoImagen": {
                "properties": {
                    "id_imagen": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_juego": {
                        "type": "integer",
                        "example": 2
                    },
                    "url": {
                        "type": "string",
                        "example": "https://example.com/imagen.jpg"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Imagen de portada del juego",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}